Moving workflows to separate packages
