+++ pstatus-notifications-workflow-ktor/src/main/kotlin/workflow/digestcounts/UploadDigestCounts.kt
@@ -1,7 +1,5 @@
 package gov.cdc.ocio.processingnotifications.workflow.digestcounts
 
-import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics
-
 
 typealias CountsByJurisdiction = Map<String, Int>
 
@@ -14,10 +12,7 @@ data class TimingMetrics(
     val max: Float,
     val mean: Float,
     val median: Float
-) {
-    val stats = DescriptiveStatistics()
-
-}
+)
 
 data class UploadDigestCounts(
     val digest: CountsByDataStreamId
