+++ pstatus-notifications-workflow-ktor/src/main/kotlin/workflow/lateuploads/NotificationWorkflow.kt
@@ -1,4 +1,4 @@
-package gov.cdc.ocio.processingnotifications.workflow
+package gov.cdc.ocio.processingnotifications.workflow.lateuploads
 
 import io.temporal.workflow.WorkflowInterface
 import io.temporal.workflow.WorkflowMethod
